requirements.txt
scipy~=1.10.1
sklearn~=0.0.post12
scikit-learn~=1.2.2
transformers~=4.26.1
torch~=2.0.0+cu118
dataclasses~=0.6
filelock~=3.12.3
numpy~=1.24.1
tqdm~=4.65.0
requests~=2.28.1
uvicorn~=0.22.0
fastapi~=0.98.0
pydantic~=1.10.9
starlette~=0.27.0
psutil~=5.9.5
packaging~=23.0
sacremoses~=0.0.53
h5py~=3.10.0
tokenizers~=0.13.2
regex~=2022.10.31
sentencepiece~=0.1.99
six~=1.16.0
nltk~=3.8.1
pillow~=9.5.0
pympler~=1.0.1
textblob~=0.17.1
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16